+++ Tests/Tests/STPFixtures.h
@@ -17,11 +17,17 @@ extern NSString *const STPTestJSONCard;
 
 extern NSString *const STPTestJSONPaymentIntent;
 
-extern NSString *const STPTestJSONSourceAlipay;
-extern NSString *const STPTestJSONSourceCard;
 extern NSString *const STPTestJSONSource3DS;
+extern NSString *const STPTestJSONSourceAlipay;
+extern NSString *const STPTestJSONSourceBancontact;
+extern NSString *const STPTestJSONSourceCard;
+extern NSString *const STPTestJSONSourceEPS;
+extern NSString *const STPTestJSONSourceGiropay;
 extern NSString *const STPTestJSONSourceiDEAL;
+extern NSString *const STPTestJSONSourceMultibanco;
+extern NSString *const STPTestJSONSourceP24;
 extern NSString *const STPTestJSONSourceSEPADebit;
+extern NSString *const STPTestJSONSourceSOFORT;
 
 @interface STPFixtures : NSObject