+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## XX.Y.Z
+* Adds `STPPaymentMethod` protocol implementation for `STPSource`. You can now call `image`/`templatedImage`/`label` on a source. [#976](https://github.com/stripe/stripe-ios/pull/976)
+
 ## 13.1.0 2018-07-13
 * Adds `STPPaymentIntent` to support PaymentIntents. [#985](https://github.com/stripe/stripe-ios/pull/985), [#986](https://github.com/stripe/stripe-ios/pull/986), [#987](https://github.com/stripe/stripe-ios/pull/987), [#988](https://github.com/stripe/stripe-ios/pull/988)
 * Reduce `NSURLSession` memory footprint. [#969](https://github.com/stripe/stripe-ios/pull/969)